requirements.txt
# Core dependencies
pyyaml>=6.0.3
pydantic>=2.5.0
jinja2>=3.1.6

# HTTP clients (for future use)
requests>=2.32.5
httpx>=0.25.0

# Utilities
python-dotenv>=1.0.0
packaging>=23.2
jsonschema>=4.25.1
